backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
networkx==3.3
python-multipart==0.0.9
pydantic==2.7.0
google-genai>=1.0.0
python-dotenv>=1.0.0
neo4j>=5.20.0
scikit-learn>=1.4.0
